Refactor NanowireNetwork into a separate file
