+++ randomnwn/dynamics.py
@@ -14,9 +14,11 @@ from numbers import Number
 from typing import Callable, Iterable
 from scipy.integrate._ivp.ivp import OdeResult
 import numpy.typing as npt
+
+from .nanowire_network import NanowireNetwork
 from .typing import *
 
-from .nanowires import get_edge_indices, NanowireNetwork
+from .nanowires import get_edge_indices
 from .calculations import solve_drain_current, solve_network
 from ._models import (
     resist_func,
@@ -24,6 +26,31 @@ from ._models import (
     _HP_model_decay,
     _HP_model_chen,
 )
+
+
+def evolve(
+    NWN: NanowireNetwork,
+    state_vars: list[str],
+    y0: npt.NDArray,
+    model: str | Callable,
+    t_eval: npt.NDArray,
+    source_node: NWNNode | list[NWNNode],
+    drain_node: NWNNode | list[NWNNode],
+    voltage_func: Callable[[npt.ArrayLike], npt.ArrayLike],
+    ivp_kwargs: dict = {},
+) -> tuple[OdeResult, list[NWNEdge]]:
+    """
+    Test
+    
+    """
+    # Get state variable derivative function
+    impl_models = {
+        "default": _HP_model_no_decay, 
+        "decay": _HP_model_decay, 
+        "chen": _HP_model_chen
+    }
+    deriv = impl_models.get(model, model)
+
 
 
 def solve_evolution(
